requirements.txt
asgiref==3.9.1
asttokens==3.0.0
attrs==25.3.0
decorator==5.2.1
Django==5.2.4
djangorestframework==3.16.0
drf-spectacular==0.28.0
executing==2.2.0
inflection==0.5.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.26.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
uritemplate==4.2.0
wcwidth==0.2.13
